apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
pytest==8.3.3
httpx==0.27.2
readability-lxml==0.8.4.1
selectolax==0.3.26
trafilatura==1.12.2
youtube-transcript-api==1.2.3
email-validator==2.2.0
pypdf==5.4.0
python-docx==1.1.2
pyinstaller==6.16.0
pymysql==1.1.2
python-tds==1.17.1
